feat(#786): NAM block editor splits params into Capture + Amp tabs
The two layers of a NAM block's controls have distinct, always-known
origins, so the editor can tab them with no per-plugin authoring: the
manifest `parameters:` axes select the capture (Capture tab), everything
else is an engine default every NAM has — input/output makeup, noise
gate, EQ, and the A2-only slim knob (Amp tab). The group of a spec IS the
editor tab, so the generic tab machinery of #780 renders them as-is.

A NAM with no capture-backed axis keeps a single group, so the editor
renders no tab bar instead of an empty Capture tab. IR/LV2 params stay
ungrouped (one flat grid).

`dispatch.rs` was over the 500-line cap, so the manifest-synthesis arms
now live one module per backend (`nam_schema`, `ir_schema`, `lv2_schema`,
shared `grid_schema`), with their tests colocated.
