+++ crates/project/tests/disk_package_schema.rs
@@ -150,6 +150,23 @@ binaries:
     // The Output knob was restored in #496 (had been dropped by #402);
     // see `nam_synthesized_schema_exposes_output_db_knob`.
     assert_eq!(amp_schema.parameters.len(), 10);
+    // #786: the editor renders one tab per group — the manifest axes pick
+    // the capture, everything else is an engine default.
+    let amp_groups: Vec<(&str, &str)> = amp_schema
+        .parameters
+        .iter()
+        .map(|p| (p.path.as_str(), p.group.as_deref().unwrap_or("")))
+        .collect();
+    for (path, group) in &amp_groups {
+        let expected = match *path {
+            "gain" | "volume" => "Capture",
+            _ => "Amp",
+        };
+        assert_eq!(
+            group, &expected,
+            "NAM param '{path}' must land in the '{expected}' tab; got: {amp_groups:?}"
+        );
+    }
 
     // ── IR: one text axis → one enum parameter ──
     let cab_schema = project::block::schema_for_block_model("cab", "ir_test_cab_e2e")
